+++ pcs_test/tier1/legacy/test_rule.py
@@ -1678,31 +1678,44 @@ class TokenPreprocessorTest(TestCase):
 
     def testNoChanges(self):
         self.assertEqual([], self.preprocessor.run([]))
+        self.assertEqual([], self.preprocessor.warning_list)
+
         self.assertEqual(
             ["#uname", "eq", "node1"],
             self.preprocessor.run(["#uname", "eq", "node1"]),
         )
+        self.assertEqual([], self.preprocessor.warning_list)
 
     def testDateSpec(self):
         self.assertEqual(["date-spec"], self.preprocessor.run(["date-spec"]))
+        self.assertEqual([], self.preprocessor.warning_list)
+
         self.assertEqual(
             ["date-spec", "hours=14"],
             self.preprocessor.run(["date-spec", "hours=14"]),
         )
+        self.assertEqual([], self.preprocessor.warning_list)
+
         self.assertEqual(
             ["date-spec", "hours weeks=6 months= moon=1"],
             self.preprocessor.run(
                 ["date-spec", "hours", "weeks=6", "months=", "moon=1"]
             ),
         )
+        self.assertEqual([], self.preprocessor.warning_list)
+
         self.assertEqual(
             ["date-spec", "foo", "hours=14"],
             self.preprocessor.run(["date-spec", "foo", "hours=14"]),
         )
+        self.assertEqual([], self.preprocessor.warning_list)
+
         self.assertEqual(
             ["date-spec", "hours=14", "foo", "hours=14"],
             self.preprocessor.run(["date-spec", "hours=14", "foo", "hours=14"]),
         )
+        self.assertEqual([], self.preprocessor.warning_list)
+
         self.assertEqual(
             [
                 "date-spec",
@@ -1724,12 +1737,16 @@ class TokenPreprocessorTest(TestCase):
                 ]
             ),
         )
+        self.assertEqual([], self.preprocessor.warning_list)
+
         self.assertEqual(
             ["#uname", "eq", "node1", "or", "date-spec", "hours=14"],
             self.preprocessor.run(
                 ["#uname", "eq", "node1", "or", "date-spec", "hours=14"]
             ),
         )
+        self.assertEqual([], self.preprocessor.warning_list)
+
         self.assertEqual(
             ["date-spec", "hours=14", "or", "#uname", "eq", "node1"],
             self.preprocessor.run(
@@ -1743,27 +1760,38 @@ class TokenPreprocessorTest(TestCase):
                 ]
             ),
         )
+        self.assertEqual([], self.preprocessor.warning_list)
 
     def testDuration(self):
         self.assertEqual(["duration"], self.preprocessor.run(["duration"]))
+        self.assertEqual([], self.preprocessor.warning_list)
+
         self.assertEqual(
             ["duration", "hours=14"],
             self.preprocessor.run(["duration", "hours=14"]),
         )
+        self.assertEqual([], self.preprocessor.warning_list)
+
         self.assertEqual(
             ["duration", "hours weeks=6 months= moon=1"],
             self.preprocessor.run(
                 ["duration", "hours", "weeks=6", "months=", "moon=1"]
             ),
         )
+        self.assertEqual([], self.preprocessor.warning_list)
+
         self.assertEqual(
             ["duration", "foo", "hours=14"],
             self.preprocessor.run(["duration", "foo", "hours=14"]),
         )
+        self.assertEqual([], self.preprocessor.warning_list)
+
         self.assertEqual(
             ["duration", "hours=14", "foo", "hours=14"],
             self.preprocessor.run(["duration", "hours=14", "foo", "hours=14"]),
         )
+        self.assertEqual([], self.preprocessor.warning_list)
+
         self.assertEqual(
             [
                 "duration",
@@ -1785,12 +1813,16 @@ class TokenPreprocessorTest(TestCase):
                 ]
             ),
         )
+        self.assertEqual([], self.preprocessor.warning_list)
+
         self.assertEqual(
             ["#uname", "eq", "node1", "or", "duration", "hours=14"],
             self.preprocessor.run(
                 ["#uname", "eq", "node1", "or", "duration", "hours=14"]
             ),
         )
+        self.assertEqual([], self.preprocessor.warning_list)
+
         self.assertEqual(
             ["duration", "hours=14", "or", "#uname", "eq", "node1"],
             self.preprocessor.run(
@@ -1804,6 +1836,7 @@ class TokenPreprocessorTest(TestCase):
                 ]
             ),
         )
+        self.assertEqual([], self.preprocessor.warning_list)
 
     def testOperationDatespec(self):
         self.assertEqual(
@@ -1813,11 +1846,27 @@ class TokenPreprocessorTest(TestCase):
             ),
         )
         self.assertEqual(
+            [
+                "Syntax 'operation=date_spec' is deprecated and will be "
+                "removed. Please use 'date-spec <date-spec options>'.",
+            ],
+            self.preprocessor.warning_list,
+        )
+
+        self.assertEqual(
             ["date-spec", "weeks=6 moon=1"],
             self.preprocessor.run(
                 ["date-spec", "weeks=6", "operation=date_spec", "moon=1"]
             ),
         )
+        self.assertEqual(
+            [
+                "Syntax 'operation=date_spec' is deprecated and will be "
+                "removed. Please use 'date-spec <date-spec options>'.",
+            ],
+            self.preprocessor.warning_list,
+        )
+
         self.assertEqual(
             ["date-spec", "weeks=6", "foo", "moon=1"],
             self.preprocessor.run(
@@ -1825,11 +1874,21 @@ class TokenPreprocessorTest(TestCase):
             ),
         )
         self.assertEqual(
+            [
+                "Syntax 'operation=date_spec' is deprecated and will be "
+                "removed. Please use 'date-spec <date-spec options>'.",
+            ],
+            self.preprocessor.warning_list,
+        )
+
+        self.assertEqual(
             ["date-spec", "weeks=6", "foo", "operation=date_spec", "moon=1"],
             self.preprocessor.run(
                 ["date-spec", "weeks=6", "foo", "operation=date_spec", "moon=1"]
             ),
         )
+        self.assertEqual([], self.preprocessor.warning_list)
+
         self.assertEqual(
             ["date-spec", "weeks=6 moon=1"],
             self.preprocessor.run(
@@ -1837,21 +1896,46 @@ class TokenPreprocessorTest(TestCase):
             ),
         )
         self.assertEqual(
+            [
+                "Syntax 'operation=date_spec' is deprecated and will be "
+                "removed. Please use 'date-spec <date-spec options>'.",
+            ],
+            self.preprocessor.warning_list,
+        )
+
+        self.assertEqual(
             ["date-spec", "weeks=6 moon=1", "foo"],
             self.preprocessor.run(
                 ["date-spec", "weeks=6", "moon=1", "operation=date_spec", "foo"]
             ),
         )
         self.assertEqual(
+            [
+                "Syntax 'operation=date_spec' is deprecated and will be "
+                "removed. Please use 'date-spec <date-spec options>'.",
+            ],
+            self.preprocessor.warning_list,
+        )
+
+        self.assertEqual(
             ["date-spec"],
             self.preprocessor.run(["date-spec", "operation=date_spec"]),
         )
+        self.assertEqual(
+            [
+                "Syntax 'operation=date_spec' is deprecated and will be "
+                "removed. Please use 'date-spec <date-spec options>'.",
+            ],
+            self.preprocessor.warning_list,
+        )
+
         self.assertEqual(
             ["date-spec", "weeks=6", "operation=foo", "moon=1"],
             self.preprocessor.run(
                 ["date-spec", "weeks=6", "operation=foo", "moon=1"]
             ),
         )
+        self.assertEqual([], self.preprocessor.warning_list)
 
     def testDateLegacySyntax(self):
         # valid syntax
@@ -1860,9 +1944,25 @@ class TokenPreprocessorTest(TestCase):
             self.preprocessor.run(["date", "start=2014-06-26", "gt"]),
         )
         self.assertEqual(
+            [
+                "Syntax 'date start=<date> gt' is deprecated and will be "
+                "removed. Please use 'date gt <date>'.",
+            ],
+            self.preprocessor.warning_list,
+        )
+
+        self.assertEqual(
             ["date", "lt", "2014-06-26"],
             self.preprocessor.run(["date", "end=2014-06-26", "lt"]),
         )
+        self.assertEqual(
+            [
+                "Syntax 'date end=<date> lt' is deprecated and will be "
+                "removed. Please use 'date lt <date>'.",
+            ],
+            self.preprocessor.warning_list,
+        )
+
         self.assertEqual(
             ["date", "in_range", "2014-06-26", "to", "2014-07-26"],
             self.preprocessor.run(
@@ -1870,10 +1970,27 @@ class TokenPreprocessorTest(TestCase):
             ),
         )
         self.assertEqual(
+            [
+                "Syntax 'date start=<date> end=<date> in_range' is deprecated "
+                "and will be removed. Please use 'date in_range <date> to "
+                "<date>'.",
+            ],
+            self.preprocessor.warning_list,
+        )
+
+        self.assertEqual(
             ["date", "in_range", "2014-06-26", "to", "2014-07-26"],
             self.preprocessor.run(
                 ["date", "end=2014-07-26", "start=2014-06-26", "in_range"]
             ),
+        )
+        self.assertEqual(
+            [
+                "Syntax 'date start=<date> end=<date> in_range' is deprecated "
+                "and will be removed. Please use 'date in_range <date> to "
+                "<date>'.",
+            ],
+            self.preprocessor.warning_list,
         )
 
         self.assertEqual(
@@ -1881,10 +1998,26 @@ class TokenPreprocessorTest(TestCase):
             self.preprocessor.run(["date", "start=2014-06-26", "gt", "foo"]),
         )
         self.assertEqual(
+            [
+                "Syntax 'date start=<date> gt' is deprecated and will be "
+                "removed. Please use 'date gt <date>'.",
+            ],
+            self.preprocessor.warning_list,
+        )
+
+        self.assertEqual(
             ["date", "lt", "2014-06-26", "foo"],
             self.preprocessor.run(["date", "end=2014-06-26", "lt", "foo"]),
         )
         self.assertEqual(
+            [
+                "Syntax 'date end=<date> lt' is deprecated and will be "
+                "removed. Please use 'date lt <date>'.",
+            ],
+            self.preprocessor.warning_list,
+        )
+
+        self.assertEqual(
             ["date", "in_range", "2014-06-26", "to", "2014-07-26", "foo"],
             self.preprocessor.run(
                 [
@@ -1897,6 +2030,15 @@ class TokenPreprocessorTest(TestCase):
             ),
         )
         self.assertEqual(
+            [
+                "Syntax 'date start=<date> end=<date> in_range' is deprecated "
+                "and will be removed. Please use 'date in_range <date> to "
+                "<date>'.",
+            ],
+            self.preprocessor.warning_list,
+        )
+
+        self.assertEqual(
             ["date", "in_range", "2014-06-26", "to", "2014-07-26", "foo"],
             self.preprocessor.run(
                 [
@@ -1907,106 +2049,164 @@ class TokenPreprocessorTest(TestCase):
                     "foo",
                 ]
             ),
+        )
+        self.assertEqual(
+            [
+                "Syntax 'date start=<date> end=<date> in_range' is deprecated "
+                "and will be removed. Please use 'date in_range <date> to "
+                "<date>'.",
+            ],
+            self.preprocessor.warning_list,
         )
 
         # invalid syntax - no change
         self.assertEqual(["date"], self.preprocessor.run(["date"]))
+        self.assertEqual([], self.preprocessor.warning_list)
+
         self.assertEqual(
             ["date", "start=2014-06-26"],
             self.preprocessor.run(["date", "start=2014-06-26"]),
         )
+        self.assertEqual([], self.preprocessor.warning_list)
+
         self.assertEqual(
             ["date", "end=2014-06-26"],
             self.preprocessor.run(["date", "end=2014-06-26"]),
         )
+        self.assertEqual([], self.preprocessor.warning_list)
+
         self.assertEqual(
             ["date", "start=2014-06-26", "end=2014-07-26"],
             self.preprocessor.run(
                 ["date", "start=2014-06-26", "end=2014-07-26"]
             ),
         )
+        self.assertEqual([], self.preprocessor.warning_list)
+
         self.assertEqual(
             ["date", "start=2014-06-26", "end=2014-07-26", "lt"],
             self.preprocessor.run(
                 ["date", "start=2014-06-26", "end=2014-07-26", "lt"]
             ),
         )
+        self.assertEqual([], self.preprocessor.warning_list)
+
         self.assertEqual(
             ["date", "start=2014-06-26", "lt", "foo"],
             self.preprocessor.run(["date", "start=2014-06-26", "lt", "foo"]),
         )
+        self.assertEqual([], self.preprocessor.warning_list)
+
         self.assertEqual(
             ["date", "start=2014-06-26", "end=2014-07-26", "gt", "foo"],
             self.preprocessor.run(
                 ["date", "start=2014-06-26", "end=2014-07-26", "gt", "foo"]
             ),
         )
+        self.assertEqual([], self.preprocessor.warning_list)
+
         self.assertEqual(
             ["date", "end=2014-06-26", "gt"],
             self.preprocessor.run(["date", "end=2014-06-26", "gt"]),
         )
+        self.assertEqual([], self.preprocessor.warning_list)
+
         self.assertEqual(
             ["date", "start=2014-06-26", "in_range", "foo"],
             self.preprocessor.run(
                 ["date", "start=2014-06-26", "in_range", "foo"]
             ),
         )
+        self.assertEqual([], self.preprocessor.warning_list)
+
         self.assertEqual(
             ["date", "end=2014-07-26", "in_range"],
             self.preprocessor.run(["date", "end=2014-07-26", "in_range"]),
         )
+        self.assertEqual([], self.preprocessor.warning_list)
+
         self.assertEqual(
             ["foo", "start=2014-06-26", "gt"],
             self.preprocessor.run(["foo", "start=2014-06-26", "gt"]),
         )
+        self.assertEqual([], self.preprocessor.warning_list)
+
         self.assertEqual(
             ["foo", "end=2014-06-26", "lt"],
             self.preprocessor.run(["foo", "end=2014-06-26", "lt"]),
         )
+        self.assertEqual([], self.preprocessor.warning_list)
 
     def testParenthesis(self):
         self.assertEqual(["("], self.preprocessor.run(["("]))
+        self.assertEqual([], self.preprocessor.warning_list)
+
         self.assertEqual([")"], self.preprocessor.run([")"]))
+        self.assertEqual([], self.preprocessor.warning_list)
+
         self.assertEqual(
             ["(", "(", ")", ")"], self.preprocessor.run(["(", "(", ")", ")"])
         )
+        self.assertEqual([], self.preprocessor.warning_list)
+
         self.assertEqual(["(", "(", ")", ")"], self.preprocessor.run(["(())"]))
+        self.assertEqual([], self.preprocessor.warning_list)
+
         self.assertEqual(
             ["a", "(", "b", ")", "c"],
             self.preprocessor.run(["a", "(", "b", ")", "c"]),
         )
+        self.assertEqual([], self.preprocessor.warning_list)
+
         self.assertEqual(
             ["a", "(", "b", "c", ")", "d"],
             self.preprocessor.run(["a", "(", "b", "c", ")", "d"]),
         )
+        self.assertEqual([], self.preprocessor.warning_list)
+
         self.assertEqual(
             ["a", ")", "b", "(", "c"],
             self.preprocessor.run(["a", ")", "b", "(", "c"]),
         )
+        self.assertEqual([], self.preprocessor.warning_list)
+
         self.assertEqual(
             ["a", "(", "b", ")", "c"], self.preprocessor.run(["a", "(b)", "c"])
         )
+        self.assertEqual([], self.preprocessor.warning_list)
+
         self.assertEqual(
             ["a", "(", "b", ")", "c"], self.preprocessor.run(["a(", "b", ")c"])
         )
+        self.assertEqual([], self.preprocessor.warning_list)
+
         self.assertEqual(
             ["a", "(", "b", ")", "c"], self.preprocessor.run(["a(b)c"])
         )
+        self.assertEqual([], self.preprocessor.warning_list)
+
         self.assertEqual(
             ["aA", "(", "bB", ")", "cC"], self.preprocessor.run(["aA(bB)cC"])
         )
+        self.assertEqual([], self.preprocessor.warning_list)
+
         self.assertEqual(
             ["(", "aA", "(", "bB", ")", "cC", ")"],
             self.preprocessor.run(["(aA(bB)cC)"]),
         )
+        self.assertEqual([], self.preprocessor.warning_list)
+
         self.assertEqual(
             ["(", "aA", "(", "(", "bB", ")", "cC", ")"],
             self.preprocessor.run(["(aA(", "(bB)cC)"]),
         )
+        self.assertEqual([], self.preprocessor.warning_list)
+
         self.assertEqual(
             ["(", "aA", "(", "(", "(", "bB", ")", "cC", ")"],
             self.preprocessor.run(["(aA(", "(", "(bB)cC)"]),
         )
+        self.assertEqual([], self.preprocessor.warning_list)
 
 
 class ExportAsExpressionTest(TestCase):
@@ -2328,6 +2528,8 @@ Location Constraints:
         )
         ac(
             output,
+            "Warning: Converting invalid score to score-attribute=pingd is "
+            "deprecated and will be removed.\n"
             "Warning: invalid score 'pingd', setting score-attribute=pingd "
             "instead\n",
         )